*.wav.gz reproducible
Hi,

I recompressed all the .wav.gz files to get rid of the embedded timestamp.
This way "lintian" won't complain about those anymore.

I know in this case, these files never changing for the same release,
the build would be reproductible, but I find it much cleaner
to do it this way than producing a +100 lines lintian-override.

This makes it easier to focus on other lintian warnings
that may be real problems that need to be solved.

Here are my steps:

$ git reset --hard
$ git clean -f
$ find . -name '*.wav'
-> no result
$ find . -name '*.wav.gz' -exec gunzip {} \;
$ find . -name '*.wav' -exec gzip -9 -n {} \;